重构项目架构，消除代码重复和冗余，将单体架构迁移到模块化架构
