+++ src/lane_identification/requirements.txt
@@ -1,0 +1,4 @@
+opencv-python==4.8.0.74
+numpy==1.24.3
+Pillow==10.0.0
+scikit-learn==1.3.0
